Updates to Main Script + New Test Script
-MAIN SCRIPT-
Reworked CPU section, now calculates average CPU usage.
Fixed bug where it would return a null instead of a 0 if CPU usage was
at 0%
Moved all writes to log file to their own section in preparation for
improving the log.
Reworked writefiles in CPU, OS, and Firewall check sections to allow
this.

-TEST SCRIPT-
Found example script that runs dxdiag and writes output to a file.
including this script in preparation of merging into the main script.