Mem Usage fix + OS test + Other Changes
Fixed Memory usage averaging (I think)
Added OS test function. At current this just prints the operating system
in the report, but it can easily be used to change script behavior based
on OS if need be.
Added small amount of new WMI calls to OS section.
Changed the way the "VolumeDIrty" value is checked and displ;ayed. There
will be no message unless the value is "True" in which case a warning
message will be displayed.