Multiple Hard drive support + save to desktop
Completely redid hard drive check section to account for multiple
logical disks. Also picks up CD drives and other disks so I'll have to
rename that section from "Harddrive info" or find a way to move the CDs
and other drive types into their own section.
Also found a very easy way to save the log to the desktop instead of the
working directory for the script. Checks the windows variable for the
desktop folder so there should be no need for any branching based on the
windows version for this.