HTML report file!!!
Completely rewrote all logfile sections to make the log file into a nice
pretty HTML report.
Changes are huge this version so I made this a new file name and new
version so that we would always have the old version just in case.
A number of places in the topmost table are blank because the variables
for them don't exist yet. But they will soon!
See bottom most comments for template on adding new sections to the
report.