CPU info + Summary Table + HD check fix + Misc.
Added WMI calls for much more relevant CPU information.
Added assorted checks and things to fill out summary table at top of
report, including the beginnings of multiple 'alert' checks to warn of
possible problems like high CPU or memory usage, or unexpected
shutdowns.
Added a check so that the report will not display information on disks
that are not currently connected to the computer.
Misc. fixes to report formatting and displayed messages, as well as
giving the report a slightly more descriptive name.