Comp Status checks + Links!
Added more checks to comp status. Also changed the way the dirty disk
check works, and updated the general formatting of the disk info
section.
Added links to any detected 'problems' that direct the view to the
relevant section. In the process also added <a> tags to the heading of
every section in the log so that new links may easily be created for new
types of detected problems added in the future.